requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# AI and ML
openai>=1.3.0
tavily-python>=0.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Vector stores and embeddings
chromadb>=0.4.18
faiss-cpu>=1.8.0
numpy>=1.26.0
scipy>=1.11.4

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Data processing
pandas==2.1.4
faker==20.1.0

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Frontend (if using Streamlit)
streamlit==1.28.2

# Logging and monitoring
loguru==0.7.2

# File processing
PyPDF2==3.0.1
python-docx==1.1.0